Move Shared interfaces into a new typings file
This will enable sharing the internal interface with addons (and catching
breakages early)
