+++ src/Interfaces.ts
@@ -3,64 +3,11 @@
  * @license MIT
  */
 
-/// <reference path="../typings/xterm.d.ts"/>
+import { IEventEmitter } from 'xterm';
+import { ITerminalOptions, ILinkMatcherOptions, IMouseZoneManager, LinkMatcherHandler, LinkMatcherValidationCallback, LineData, IColorSet, IRenderer, IBufferSet, IBuffer, ISelectionManager } from '../typings/xterm-internal';
+import { ICharset } from './Interfaces';
 
-import { Terminal as PublicTerminal, ITerminalOptions as IPublicTerminalOptions, IEventEmitter } from 'xterm';
-import { ICharset, ILinkMatcherOptions } from './Interfaces';
-import { LinkMatcherHandler, LinkMatcherValidationCallback, LineData } from './Types';
-import { IColorSet, IRenderer } from './renderer/Interfaces';
-import { IMouseZoneManager } from './input/Interfaces';
 
-export interface IBrowser {
-  isNode: boolean;
-  userAgent: string;
-  platform: string;
-  isFirefox: boolean;
-  isMSIE: boolean;
-  isMac: boolean;
-  isIpad: boolean;
-  isIphone: boolean;
-  isMSWindows: boolean;
-}
-
-export interface IBufferAccessor {
-  buffer: IBuffer;
-}
-
-export interface IElementAccessor {
-  element: HTMLElement;
-}
-
-export interface ILinkifierAccessor {
-  linkifier: ILinkifier;
-}
-
-export interface ITerminal extends PublicTerminal, ILinkifierAccessor, IBufferAccessor, IElementAccessor {
-  selectionManager: ISelectionManager;
-  charMeasure: ICharMeasure;
-  renderer: IRenderer;
-  browser: IBrowser;
-  writeBuffer: string[];
-  cursorHidden: boolean;
-  cursorState: number;
-  defAttr: number;
-  options: ITerminalOptions;
-  buffers: IBufferSet;
-  isFocused: boolean;
-  mouseHelper: IMouseHelper;
-  bracketedPasteMode: boolean;
-
-  /**
-   * Emit the 'data' event and populate the given data.
-   * @param data The data to populate in the event.
-   */
-  handler(data: string): void;
-  scrollLines(disp: number, suppressScrollEvent?: boolean): void;
-  cancel(ev: Event, force?: boolean): boolean | void;
-  log(text: string): void;
-  showCursor(): void;
-  blankLine(cur?: boolean, isWrapped?: boolean, cols?: number): LineData;
-}
 
 /**
  * This interface encapsulates everything needed from the Terminal by the
@@ -125,48 +72,6 @@ export interface IInputHandlingTerminal extends IEventEmitter {
   setOption(key: string, value: any): void;
 }
 
-// TODO: The options that are not in the public API should be reviewed
-export interface ITerminalOptions extends IPublicTerminalOptions {
-  cancelEvents?: boolean;
-  convertEol?: boolean;
-  debug?: boolean;
-  handler?: (data: string) => void;
-  screenKeys?: boolean;
-  termName?: string;
-  useFlowControl?: boolean;
-}
-
-export interface IBuffer {
-  lines: ICircularList<LineData>;
-  ydisp: number;
-  ybase: number;
-  y: number;
-  x: number;
-  tabs: any;
-  scrollBottom: number;
-  scrollTop: number;
-  savedY: number;
-  savedX: number;
-  isCursorInViewport: boolean;
-  translateBufferLineToString(lineIndex: number, trimRight: boolean, startCol?: number, endCol?: number): string;
-  nextStop(x?: number): number;
-  prevStop(x?: number): number;
-}
-
-export interface IBufferSet {
-  alt: IBuffer;
-  normal: IBuffer;
-  active: IBuffer;
-
-  activateNormalBuffer(): void;
-  activateAltBuffer(): void;
-}
-
-export interface IMouseHelper {
-  getCoords(event: {pageX: number, pageY: number}, element: HTMLElement, charMeasure: ICharMeasure, lineHeight: number, colCount: number, rowCount: number, isSelection?: boolean): [number, number];
-  getRawByteCoords(event: MouseEvent, element: HTMLElement, charMeasure: ICharMeasure, lineHeight: number, colCount: number, rowCount: number): { x: number, y: number };
-}
-
 export interface IViewport {
   syncScrollArea(): void;
   onWheel(ev: WheelEvent): void;
@@ -175,79 +80,12 @@ export interface IViewport {
   onThemeChanged(colors: IColorSet): void;
 }
 
-export interface ISelectionManager {
-  selectionText: string;
-  selectionStart: [number, number];
-  selectionEnd: [number, number];
-
-  disable(): void;
-  enable(): void;
-  setBuffer(buffer: IBuffer): void;
-  setSelection(row: number, col: number, length: number): void;
-}
-
 export interface ICompositionHelper {
   compositionstart(): void;
   compositionupdate(ev: CompositionEvent): void;
   compositionend(): void;
   updateCompositionElements(dontRecurse?: boolean): void;
   keydown(ev: KeyboardEvent): boolean;
-}
-
-export interface ICharMeasure {
-  width: number;
-  height: number;
-  measure(options: ITerminalOptions): void;
-}
-
-export interface ILinkifier extends IEventEmitter {
-  attachToDom(mouseZoneManager: IMouseZoneManager): void;
-  linkifyRows(start: number, end: number): void;
-  setHypertextLinkHandler(handler: LinkMatcherHandler): void;
-  setHypertextValidationCallback(callback: LinkMatcherValidationCallback): void;
-  registerLinkMatcher(regex: RegExp, handler: LinkMatcherHandler, options?: ILinkMatcherOptions): number;
-  deregisterLinkMatcher(matcherId: number): boolean;
-}
-
-export interface ICircularList<T> extends IEventEmitter {
-  length: number;
-  maxLength: number;
-  forEach: (callbackfn: (value: T, index: number) => void) => void;
-
-  get(index: number): T;
-  set(index: number, value: T): void;
-  push(value: T): void;
-  pop(): T;
-  splice(start: number, deleteCount: number, ...items: T[]): void;
-  trimStart(count: number): void;
-  shiftElements(start: number, count: number, offset: number): void;
-}
-
-export interface ILinkMatcherOptions {
-  /**
-   * The index of the link from the regex.match(text) call. This defaults to 0
-   * (for regular expressions without capture groups).
-   */
-  matchIndex?: number;
-  /**
-   * A callback that validates an individual link, returning true if valid and
-   * false if invalid.
-   */
-  validationCallback?: LinkMatcherValidationCallback;
-  /**
-   * A callback that fires when the mouse hovers over a link.
-   */
-  tooltipCallback?: LinkMatcherHandler;
-  /**
-   * A callback that fires when the mouse leaves a link that was hovered.
-   */
-  leaveCallback?: () => void;
-  /**
-   * The priority of the link matcher, this defines the order in which the link
-   * matcher is evaluated relative to others, from highest to lowest. The
-   * default value is 0.
-   */
-  priority?: number;
 }
 
 /**
@@ -300,30 +138,6 @@ export interface IInputHandler {
   /** CSI r */ setScrollRegion(params?: number[]): void;
   /** CSI s */ saveCursor(params?: number[]): void;
   /** CSI u */ restoreCursor(params?: number[]): void;
-}
-
-export interface ITheme {
-  foreground?: string;
-  background?: string;
-  cursor?: string;
-  cursorAccent?: string;
-  selection?: string;
-  black?: string;
-  red?: string;
-  green?: string;
-  yellow?: string;
-  blue?: string;
-  magenta?: string;
-  cyan?: string;
-  white?: string;
-  brightBlack?: string;
-  brightRed?: string;
-  brightGreen?: string;
-  brightYellow?: string;
-  brightBlue?: string;
-  brightMagenta?: string;
-  brightCyan?: string;
-  brightWhite?: string;
 }
 
 export interface ILinkMatcher {
