+++ src/renderer/Interfaces.ts
@@ -3,25 +3,9 @@
  * @license MIT
  */
 
-import { ITerminal, ITerminalOptions, ITheme } from '../Interfaces';
+import { IColorSet, IRenderDimensions, ITerminal } from '../../typings/xterm-internal';
+
 import { IEventEmitter } from 'xterm';
-
-export interface IRenderer extends IEventEmitter {
-  dimensions: IRenderDimensions;
-  colorManager: IColorManager;
-
-  setTheme(theme: ITheme): IColorSet;
-  onWindowResize(devicePixelRatio: number): void;
-  onResize(cols: number, rows: number, didCharSizeChange: boolean): void;
-  onCharSizeChanged(): void;
-  onBlur(): void;
-  onFocus(): void;
-  onSelectionChanged(start: [number, number], end: [number, number]): void;
-  onCursorMove(): void;
-  onOptionsChanged(): void;
-  clear(): void;
-  queueRefresh(start: number, end: number): void;
-}
 
 export interface IRenderLayer {
   /**
@@ -69,32 +53,4 @@ export interface IRenderLayer {
    * Clear the state of the render layer.
    */
   reset(terminal: ITerminal): void;
-}
-
-export interface IColorManager {
-  colors: IColorSet;
-}
-
-export interface IColorSet {
-  foreground: string;
-  background: string;
-  cursor: string;
-  cursorAccent: string;
-  selection: string;
-  ansi: string[];
-}
-
-export interface IRenderDimensions {
-  scaledCharWidth: number;
-  scaledCharHeight: number;
-  scaledCellWidth: number;
-  scaledCellHeight: number;
-  scaledCharLeft: number;
-  scaledCharTop: number;
-  scaledCanvasWidth: number;
-  scaledCanvasHeight: number;
-  canvasWidth: number;
-  canvasHeight: number;
-  actualCellWidth: number;
-  actualCellHeight: number;
 }
